requirements.txt
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.4.0
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python==4.10.0.84
numpy>=1.26.0
